backend/innstal/requirements.txt
Django==1.11
psycopg2==2.7.3.2
Pillow==5.0.0
django-ckeditor==5.4.0
djangorestframework==3.7.7